requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pillow>=10.0.0
pytesseract>=0.3.10
httpx>=0.27.0
rapidfuzz>=3.6.0
